finish refactor code
