+++ app/models/covid_data_model.py
@@ -1,0 +1,5 @@
+import httpx
+
+covid_data_JSON = {}
+with httpx.Client() as client:
+    covid_data_JSON = client.get("https://data.covid19.go.id/public/api/update.json").json()
