+++ app/utils/constant.py
@@ -1,4 +1,9 @@
+import httpx
 from datetime import date
+
+covid_data_JSON = {}
+with httpx.Client() as client:
+    covid_data_JSON = client.get("https://data.covid19.go.id/public/api/update.json").json()
 
 EARLIEST_YEAR_CASE = "2020"
 EARLIEST_MONTH_CASE = "03"
